requirements.txt
streamlit

google-generativeai
supabase
pandas
pytz
plotly
numpy
matplotlib
openpyxl
xlsxwriter
fpdf